Patch แก้ไขอนิเมชันของประกอบเกมทั้งหมด ถ้ำ บอนไฟ หญ้า
แก้ไขอนิเมชันของประกอบฉาก ทั้งหมด เห็ด หญ้า ถ้ำ บอนไฟ
เพิ่มTotemที่เอาไว้้Spawnขวดยาเลือดเพิ่มHp25ทุกๆ20วิ
ตอนนี้ตกน้ำขึ้นมาไม่ได้ เมื่อถึงความลึกที่กำหนดจะตายห่าทันที

+++ Assets/Scripts/playerStat.cs
@@ -14,8 +14,8 @@ public class playerStat : MonoBehaviour
     //hp&sp test
     public int damage;
     //public float sp;
-    public int MaxHp = 100;
-    public int currentHp;   
+    public float MaxHp = 100;
+    public float currentHp;   
     public int PlayerSoul;
     public int PlayerCurrentSoul = 1;
     public GameObject Player;
